+++ bundled/tool/install_dependencies.py
@@ -3,6 +3,15 @@ import sys
 from pathlib import Path
 
 def install_dependencies(extension_root_dir):
+    """
+    Install dependencies required for the Kedro extension.
+
+    Args:
+        extension_root_dir (str): The root directory of the extension.
+    Raises:
+        ImportError: If the required dependencies are not found.
+    """
+    ...
     libs_path = Path(extension_root_dir) / "bundled" / "libs"
     requirements_path = Path(extension_root_dir) / "kedro-viz-requirements.txt"
